backend/requirements.txt
fastapi==0.116.1
uvicorn==0.35.0
python-dotenv==1.1.1
motor==3.7.1
PyPDF2==3.0.1
python-multipart==0.0.20
langchain-core==0.3.72
langchain-google-genai==2.1.8
langgraph==0.5.4
langgraph-api==0.2.102
langgraph-checkpoint==2.1.1
langgraph-cli==0.3.6
langgraph-prebuilt==0.5.2
langgraph-runtime-inmem==0.6.1
langgraph-sdk==0.1.74
pydantic==2.11.7
pydantic-core==2.33.2
requests==2.32.4
reportlab==4.4.3
structlog==25.4.0
tenacity==9.1.2
typing-extensions==4.14.1